@@ -146,85 +147,79 @@
 			]
 		}
 	},
-	"scripts": {
-		"vscode:prepublish": "npm run package",
-		"compile": "npm run check-types && npm run lint && node esbuild.js",
-		"watch": "npm-run-all -p watch:*",
-		"watch:esbuild": "node esbuild.js --watch",
-		"watch:tsc": "tsc --noEmit --watch --project tsconfig.json",
-		"package": "npm run build:webview && npm run check-types && npm run lint && node esbuild.js --production",
-		"compile-tests": "tsc -p . --outDir out",
-		"watch-tests": "tsc -p . -w --outDir out",
-		"pretest": "npm run compile-tests && npm run compile && npm run lint",
-		"check-types": "tsc --noEmit",
-		"lint": "eslint src --ext ts",
-		"test": "vscode-test",
-		"install:all": "npm install && cd webview-ui-vite && npm install",
-		"start:webview": "cd webview-ui-vite && npm run dev",
-		"build:webview": "cd webview-ui-vite && npm run build",
-		"test:webview": "cd webview-ui-vite && npm run test",
-		"postinstall": "npm rebuild sharp"
+ "scripts": {
+  "vscode:prepublish": "pnpm run package",
+  "compile": "pnpm run check-types && pnpm run lint && tsx esbuild.ts",
+  "watch": "pnpm run-all -p watch:*",
+  "watch:esbuild": "tsx esbuild.ts --watch",
+  "watch:tsc": "tsc --noEmit --watch --project tsconfig.json",
+  "package": "pnpm run build:webview && pnpm run check-types && pnpm run lint && tsx esbuild.ts --production",
+  "pretest": "pnpm run lint",
+  "check-types": "tsc --noEmit",
+  "lint": "eslint src",
+  "test": "vscode-test",
+		"install:all": "pnpm install && cd webview-ui-vite && pnpm install",
+		"start:webview": "cd webview-ui-vite && pnpm run dev",
+		"build:webview": "cd webview-ui-vite && pnpm run build",
+		"test:webview": "cd webview-ui-vite && pnpm run test"
 	},
 	"devDependencies": {
 		"@types/diff": "^5.2.3",
 		"@types/diff-match-patch": "^1.0.36",
-		"@types/jest": "^29.5.13",
-		"@types/lodash": "^4.17.9",
-		"@types/mocha": "^10.0.7",
-		"@types/node": "20.x",
-		"@types/string-similarity": "^4.0.2",
-		"@types/vscode": "^1.82.0",
-		"@typescript-eslint/eslint-plugin": "^7.14.1",
-		"@typescript-eslint/parser": "^7.11.0",
-		"@vitejs/plugin-react-swc": "^3.7.0",
+		"@types/jest": "^29.5.14",
+		"@types/lodash": "^4.17.13",
+		"@types/mocha": "^10.0.10",
+		"@types/node": "~20.17.7",
+		"@types/vscode": "^1.95.0",
+		"@typescript-eslint/eslint-plugin": "^8.15.0",
+		"@typescript-eslint/parser": "^8.15.0",
+		"@vitejs/plugin-react-swc": "^3.7.1",
 		"@vscode/test-cli": "^0.0.9",
 		"@vscode/test-electron": "^2.4.1",
 		"esbuild": "^0.21.5",
-		"eslint": "^8.57.0",
-		"npm-run-all": "^4.1.5",
-		"typescript": "^5.4.5"
+		"eslint": "^9.15.0",
+		"@jest/globals": "^29.7.0",
+		"typescript": "^5.7.2"
 	},
 	"dependencies": {
-		"@amplitude/analytics-browser": "^2.11.2",
+		"@amplitude/analytics-browser": "^2.11.9",
 		"@amplitude/analytics-node": "^1.3.6",
-		"@anthropic-ai/bedrock-sdk": "^0.10.2",
-		"@anthropic-ai/sdk": "^0.26.0",
-		"@anthropic-ai/vertex-sdk": "^0.4.1",
+		"@anthropic-ai/bedrock-sdk": "^0.10.4",
+		"@anthropic-ai/sdk": "^0.26.1",
+		"@anthropic-ai/vertex-sdk": "^0.4.3",
 		"@ewoudenberg/difflib": "^0.1.0",
 		"@types/clone-deep": "^4.0.4",
 		"@types/pdf-parse": "^1.1.4",
 		"@vscode/codicons": "^0.0.36",
 		"ansi-regex": "^6.1.0",
-		"axios": "^1.7.4",
+		"axios": "^1.7.7",
 		"clone-deep": "^4.0.1",
 		"default-shell": "^2.2.0",
 		"delay": "^6.0.0",
 		"diff": "^5.2.0",
 		"diff-match-patch": "^1.0.5",
 		"dotenv": "^16.4.5",
-		"execa": "^9.3.0",
+		"execa": "^9.5.1",
 		"globby": "^14.0.2",
 		"image-size": "^1.1.1",
-		"isbinaryfile": "^5.0.2",
+		"isbinaryfile": "^5.0.4",
 		"jest": "^29.7.0",
-		"jotai": "^2.9.3",
-		"loadash": "^1.0.0",
+		"jotai": "^2.10.3",
 		"lodash": "^4.17.21",
 		"mammoth": "^1.8.0",
-		"nanoid": "^5.0.7",
-		"openai": "^4.54.0",
+		"nanoid": "^5.0.8",
+		"openai": "^4.73.0",
 		"os-name": "^6.0.0",
 		"p-queue": "^8.0.1",
 		"p-wait-for": "^5.0.2",
 		"pdf-parse": "^1.1.1",
 		"puppeteer-chromium-resolver": "^23.0.0",
-		"puppeteer-core": "^23.6.0",
+		"puppeteer-core": "^23.9.0",
 		"rxjs": "^7.8.1",
 		"serialize-error": "^11.0.3",
-		"sharp": "^0.33.5",
-		"string-similarity": "^4.0.4",
 		"tree-kill": "^1.2.2",
-		"tree-sitter-wasms": "^0.1.11",
+		"tree-sitter-wasms": "^0.1.12",
+		"tsx": "^4.19.2",
 		"web-tree-sitter": "^0.22.6",
 		"zod": "^3.23.8"
 	}
